Use waitress as production server
