+++ imjoy_elfinder/templates/elfinder/elfinder.html
@@ -6,6 +6,7 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=2">
 		<title>ImJoy elFinder v{{ IMJOY_ELFINDER_VERSION }}</title>
 		<script src="https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.6/require.min.js"></script>
+		<script src="https://lib.imjoy.io/imjoy-loader.js"></script>
 		<script>{% include "elfinder/main.js" %}</script>
 	</head>
 	<body>
