+++ requirements.txt
@@ -5,3 +5,4 @@ pyramid
 pyramid_jinja2
 typing_extensions
 werkzeug
+waitress
